Create routes.rs
